feat(tui): redesign BLOCKS view layout (header, sub-tabs, sidebar, nav)
